requirements.txt
fastapi
uvicorn
streamlit
pillow
torch
torchvision
python-multipart
requests
